Improve logic
