+++ frontend/sun-shield/package.json
@@ -13,6 +13,7 @@
     "@types/node": "^16.18.126",
     "@types/react": "^18.2.65",
     "@types/react-dom": "^18.2.17",
+    "axios": "^1.8.3",
     "mobx": "^6.13.6",
     "mobx-react": "^9.2.0",
     "mobx-react-lite": "^4.1.0",
